requirements.txt
beautifulsoup4==4.14.2
fpdf2==2.8.4
pandas==2.3.3
pdfplumber==0.11.7
Pillow==11.3.0
Requests==2.32.5
streamlit==1.50.0
openpyxl==3.1.5
dotenv==0.9.9